Add support for non-smooth dynamic wallpapers
This commit draws a line between "smooth" and "non-smooth" dynamic
wallpapers. The difference between the two lies in how this plugin
computes blend factors between the bottom and the top layer.

With smooth dynamic wallpapers, the blend factor goes from 0 to 1.
However, with non-smooth dynamic wallpapers, the blend factor is
either 0 or 1.

The reason why one would want that is to avoid visual artifacts if
the dynamic wallpaper contains a few moving objects.