Work around the lack of onValueModified signal in QQC 1
SpinBox in QtQuick Controls 2 has a signal that is emitted when the
value is changed explicitly by the user. In QtQuick Controls 1, there is
no such a signal.

The most straightforward thing would be to switch to QQC2, but the
problem is that the SpinBox API in there is extremely stripped down. You
will need to re-invent the wheel to get basic functionality.

This fixes a bug where the update interval is reset to 1 unexpectedly.